Fix includ file pats as result of move
