+++ opencog/cogserver/modules/RuleEngineModule.h
@@ -24,7 +24,7 @@
 #ifndef RULEENGINEMODULE_H_
 #define RULEENGINEMODULE_H_
 
-#include <opencog/server/Module.h>
+#include <opencog/cogserver/server/Module.h>
 #include <opencog/rule-engine/InferenceSCM.h>
 
 namespace opencog {
