+++ opencog/cogserver/modules/SuRealModule.h
@@ -24,7 +24,7 @@
 #ifndef _OPENCOG_SUREAL_MODULE_H
 #define _OPENCOG_SUREAL_MODULE_H
 
-#include <opencog/server/Module.h>
+#include <opencog/cogserver/server/Module.h>
 #include <opencog/nlp/sureal/SuRealSCM.h>
 
 namespace opencog
