+++ opencog/cogserver/server/Agent.cc
@@ -1,5 +1,5 @@
 /*
- * opencog/server/Agent.cc
+ * opencog/cogserver/server/Agent.cc
  *
  * Copyright (C) 2008 by OpenCog Foundation
  * All Rights Reserved
@@ -24,7 +24,7 @@
 
 #include "Agent.h"
 
-#include <opencog/server/CogServer.h>
+#include <opencog/cogserver/server/CogServer.h>
 #include <opencog/util/Config.h>
 
 #define DEBUG
