+++ opencog/cogserver/server/ListModulesRequest.cc
@@ -1,5 +1,5 @@
 #include "ListModulesRequest.h"
-#include <opencog/server/CogServer.h>
+#include <opencog/cogserver/server/CogServer.h>
 
 using namespace opencog;
