+++ opencog/cogserver/server/ListRequest.cc
@@ -1,5 +1,5 @@
 /*
- * opencog/server/ListRequest.cc
+ * opencog/cogserver/server/ListRequest.cc
  *
  * Copyright (C) 2008 by OpenCog Foundation
  * All Rights Reserved
@@ -27,7 +27,7 @@
 #include <opencog/atomspace/AtomSpace.h>
 #include <opencog/atomspace/ClassServer.h>
 #include <opencog/atomspace/types.h>
-#include <opencog/server/CogServer.h>
+#include <opencog/cogserver/server/CogServer.h>
 
 using namespace opencog;
