+++ opencog/cogserver/shell/PythonShell.cc
@@ -32,7 +32,7 @@
 #ifdef HAVE_CYTHON
 
 #include <opencog/cython/PythonEval.h>
-#include <opencog/server/ConsoleSocket.h>
+#include <opencog/cogserver/server/ConsoleSocket.h>
 #include "PythonShell.h"
 
 using namespace opencog;
