+++ opencog/cogserver/shell/PythonShellModule.cc
@@ -32,7 +32,7 @@
 #ifdef HAVE_CYTHON
 
 #include <opencog/cython/PythonEval.h>
-#include <opencog/server/ConsoleSocket.h>
+#include <opencog/cogserver/server/ConsoleSocket.h>
 #include <opencog/util/Logger.h>
 #include <opencog/util/platform.h>
 #include "PythonShellModule.h"
